app/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyodbc==4.0.35
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0
